Renamed files with algorithms to match the names in the paper.
