+++ src/global.h
