+++ src/main.cpp
@@ -1,7 +1,7 @@
-#include "greedy_ac.h"
-#include "greedy.h"
-#include "generalized_simplitigs.h"
-#include "generalized_simplitigs_ac.h"
+#include "global_ac.h"
+#include "global.h"
+#include "local.h"
+#include "local_ac.h"
 #include "parser.h"
 #include "streaming.h"
 #include "output.h"
