+++ src/unittest.cpp
@@ -1,9 +1,9 @@
 #include "parser_unittest.h"
 #include "kmers_unittest.h"
-#include "greedy_unittest.h"
-#include "greedy_unittest_ac.h"
-#include "generalized_simplitigs_unittest.h"
-#include "generalized_simplitigs_ac_unittest.h"
+#include "global_unittest.h"
+#include "global_ac_unittes.h"
+#include "local_unittest.h"
+#include "local_ac_unittest.h"
 #include "streaming_unittest.h"
 
 #include "gtest/gtest.h"
